test_requirements.txt
django-cms~=4.1
djangocms-alias
djangocms-versioning==2.3.2
djangocms-aldryn-categories
djangocms-aldryn-people
djangocms-text
aldryn-apphooks-config
django-haystack
django-taggit

# NOTE: These are for Filer
django-mptt
django_polymorphic

djangocms-aldryn-search
coverage>=3.7.1
dj-database-url
django-classy-tags
django-sekizai
django-treebeard>=2.0
djangocms-admin-style
django-app-helper
djangocms-text-ckeditor
docopt
flake8
tox>=2.9.1
pyflakes
selenium
python-dateutil
pytz
reportlab
svglib

# docs
sphinx
sphinx-rtd-theme

sphinxcontrib-spelling
pyenchant
